chore(cstor-volume-mgmt): Convert grpc and volume-mgmt binaries into a single binary.
The cstor-volume-mgmt pod executes two different binaries for GRPC
server and CstorVolume watcher.
This was done earlier so that parallel progress can be made on GRPC
code, now this is not required.

Moved all the logic of cstor-volume-grpc into cstor-volume-mgmt.
Removed CLI commands of cstor-volume-grpc(which were being used during development for manual
testing).
Moved code in cmd/cstor-volume-grpc to pkg/grpc and import it for reuse.

Updated build.sh to use grpc build into volume-mgmt.
Updated buildscripts/cstor-volume-mgmt/entrypoint.sh to use single
binary.
