+++ pkg/grpc/v1alpha1/api/handler_test.go
@@ -3,7 +3,7 @@ package api
 import (
 	"testing"
 
-	"github.com/openebs/maya/pkg/client/generated/cstor-volume-grpc/v1alpha1"
+	"github.com/openebs/maya/pkg/client/generated/cstor-volume-mgmt/v1alpha1"
 )
 
 type TestUnixSock struct{}
